Restructured dronetraffic plugin
- Renamed to dronetraf.py
- Dronetraf now implemented as an Entity, and thus inherits additional functionality from simulator
- Maintains an up-to-date list of all active drones
